+++ internal/sdkprovider/service/valkey/valkey_user.go
@@ -1,0 +1,192 @@
+package valkey
+
+import (
+	"context"
+
+	avngen "github.com/aiven/go-client-codegen"
+	"github.com/aiven/go-client-codegen/handler/serviceuser"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+	"github.com/aiven/terraform-provider-aiven/internal/common"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil"
+	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig"
+)
+
+var aivenValkeyUserSchema = map[string]*schema.Schema{
+	"project":      schemautil.CommonSchemaProjectReference,
+	"service_name": schemautil.CommonSchemaServiceNameReference,
+
+	"username": {
+		Type:         schema.TypeString,
+		Required:     true,
+		ForceNew:     true,
+		ValidateFunc: schemautil.GetServiceUserValidateFunc(),
+		Description:  userconfig.Desc("The actual name of the Valkey User.").ForceNew().Referenced().Build(),
+	},
+	"password": {
+		Type:             schema.TypeString,
+		Optional:         true,
+		Sensitive:        true,
+		Computed:         true,
+		DiffSuppressFunc: schemautil.EmptyObjectDiffSuppressFunc,
+		Description:      "The password of the Valkey User.",
+	},
+	"valkey_acl_categories": {
+		Type:         schema.TypeList,
+		Optional:     true,
+		ForceNew:     true,
+		RequiredWith: []string{"valkey_acl_commands", "valkey_acl_keys"},
+		Description:  userconfig.Desc("Defines command category rules.").RequiredWith("valkey_acl_commands", "valkey_acl_keys").ForceNew().Build(),
+		Elem: &schema.Schema{
+			Type: schema.TypeString,
+		},
+	},
+	"valkey_acl_commands": {
+		Type:         schema.TypeList,
+		Optional:     true,
+		ForceNew:     true,
+		RequiredWith: []string{"valkey_acl_categories", "valkey_acl_keys"},
+		Description:  userconfig.Desc("Defines rules for individual commands.").RequiredWith("valkey_acl_categories", "valkey_acl_keys").ForceNew().Build(),
+		Elem: &schema.Schema{
+			Type: schema.TypeString,
+		},
+	},
+	"valkey_acl_keys": {
+		Type:         schema.TypeList,
+		Optional:     true,
+		ForceNew:     true,
+		RequiredWith: []string{"valkey_acl_categories", "valkey_acl_commands"},
+		Description:  userconfig.Desc("Defines key access rules.").RequiredWith("valkey_acl_categories", "valkey_acl_keys").ForceNew().Build(),
+		Elem: &schema.Schema{
+			Type: schema.TypeString,
+		},
+	},
+	"valkey_acl_channels": {
+		Type:        schema.TypeList,
+		Optional:    true,
+		ForceNew:    true,
+		Description: userconfig.Desc("Defines the permitted pub/sub channel patterns.").ForceNew().Build(),
+		Elem: &schema.Schema{
+			Type: schema.TypeString,
+		},
+	},
+
+	// computed fields
+	"type": {
+		Type:        schema.TypeString,
+		Computed:    true,
+		Description: "Type of the user account. Tells whether the user is the primary account or a regular account.",
+	},
+}
+
+func ResourceValkeyUser() *schema.Resource {
+	return &schema.Resource{
+		Description: userconfig.Desc(
+			"Creates and manages an " +
+				"[Aiven for Valkey](https://aiven.io/docs/products/valkey) user.",
+		).
+			AvailabilityType(userconfig.Beta).
+			Build(),
+		CreateContext: common.WithGenClient(resourceValkeyUserCreate),
+		UpdateContext: common.WithGenClient(resourceValkeyUserUpdate),
+		ReadContext:   common.WithGenClient(resourceValkeyUserRead),
+		DeleteContext: schemautil.ResourceServiceUserDelete,
+		Importer: &schema.ResourceImporter{
+			StateContext: schema.ImportStatePassthroughContext,
+		},
+		Timeouts: schemautil.DefaultResourceTimeouts(),
+
+		Schema: aivenValkeyUserSchema,
+	}
+}
+
+func resourceValkeyUserCreate(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
+	projectName := d.Get("project").(string)
+	serviceName := d.Get("service_name").(string)
+	username := d.Get("username").(string)
+	categories := schemautil.FlattenToString(d.Get("valkey_acl_categories").([]interface{}))
+	commands := schemautil.FlattenToString(d.Get("valkey_acl_commands").([]interface{}))
+	keys := schemautil.FlattenToString(d.Get("valkey_acl_keys").([]interface{}))
+	channels := schemautil.FlattenToString(d.Get("valkey_acl_channels").([]interface{}))
+	var req = serviceuser.ServiceUserCreateIn{
+		Username: username,
+		AccessControl: &serviceuser.AccessControlIn{
+			ValkeyAclCategories: &categories,
+			ValkeyAclCommands:   &commands,
+			ValkeyAclKeys:       &keys,
+			ValkeyAclChannels:   &channels,
+		},
+	}
+
+	_, err := client.ServiceUserCreate(
+		ctx,
+		projectName,
+		serviceName,
+		&req,
+	)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	if _, ok := d.GetOk("password"); ok {
+		var req = serviceuser.ServiceUserCredentialsModifyIn{NewPassword: schemautil.OptionalStringPointer(d, "password"),
+			Operation: serviceuser.OperationTypeResetCredentials}
+		_, err := client.ServiceUserCredentialsModify(ctx, projectName, serviceName, username, &req)
+		if err != nil {
+			return diag.FromErr(err)
+		}
+	}
+
+	d.SetId(schemautil.BuildResourceID(projectName, serviceName, username))
+
+	return resourceValkeyUserRead(ctx, d, client)
+}
+
+func resourceValkeyUserUpdate(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
+	projectName, serviceName, username, err := schemautil.SplitResourceID3(d.Id())
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	_, err = client.ServiceUserCredentialsModify(ctx, projectName, serviceName, username, &serviceuser.ServiceUserCredentialsModifyIn{
+		NewPassword: schemautil.OptionalStringPointer(d, "password"),
+	})
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	return resourceValkeyUserRead(ctx, d, client)
+}
+
+func resourceValkeyUserRead(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
+	projectName, serviceName, username, err := schemautil.SplitResourceID3(d.Id())
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	user, err := client.ServiceUserGet(ctx, projectName, serviceName, username)
+	if err != nil {
+		return diag.FromErr(schemautil.ResourceReadHandleNotFound(err, d))
+	}
+
+	err = schemautil.CopyServiceUserGenPropertiesFromAPIResponseToTerraform(d, user, projectName, serviceName)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	if err := d.Set("valkey_acl_keys", user.AccessControl.ValkeyAclKeys); err != nil {
+		return diag.FromErr(err)
+	}
+	if err := d.Set("valkey_acl_categories", user.AccessControl.ValkeyAclCategories); err != nil {
+		return diag.FromErr(err)
+	}
+	if err := d.Set("valkey_acl_commands", user.AccessControl.ValkeyAclCommands); err != nil {
+		return diag.FromErr(err)
+	}
+	if err := d.Set("valkey_acl_channels", user.AccessControl.ValkeyAclChannels); err != nil {
+		return diag.FromErr(err)
+	}
+
+	return nil
+}